Restructured as multi-project build
