+++ plugin-tests/src/test/scala/com/olegpy/bm4/TestNoTuples.scala
