+++ plugin/src/main/resources/scalac-plugin.xml
