+++ plugin/src/main/scala/com/olegpy/bm4/NoMapIdentity.scala
