+++ plugin/src/main/scala/com/olegpy/bm4/TreeUtils.scala
